oai_code_evaluator/pyproject.toml
[tool.poetry]
name = "oai-code-evaluator"
version = "0.3.0"
description = "Herramienta CLI para simular el rol de revisor: evalúa prompts, calificaciones y correcciones de código/model responses siguiendo rúbricas LATAM Outlier.ai"
authors = ["Carlos <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "oai_code_evaluator"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
rich = "^13.7.1"
pyyaml = "^6.0.2"
jsonschema = "^4.23.0"
jinja2 = "^3.1.4"
# Se mantiene liviano; ampliar según necesidades reales (por ejemplo para ejecutar código en sandboxes)

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[tool.poetry.scripts]
oai-eval = "oai_code_evaluator.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
